Move from utils to pipeline
