+++ heareval/tasks/pipeline.py
@@ -3,15 +3,80 @@ Generic pipelines for datasets
 """
 
 import os
+import json
+import shutil
+from glob import glob
 from pathlib import Path
 from typing import Dict, List, Union
 from urllib.parse import urlparse
-from slugify import slugify
 
 import luigi
 import pandas as pd
+from tqdm import tqdm
+from slugify import slugify
 
-import heareval.tasks.util.luigi as luigi_util
+from heareval.tasks.util.luigi import (
+    WorkTask,
+    download_file,
+    filename_to_int_hash,
+    new_basedir,
+)
+
+
+class DownloadCorpus(WorkTask):
+    """
+    Downloads from the url and saveds it in the workdir with name
+    outfile
+    """
+
+    url = luigi.Parameter()
+    outfile = luigi.Parameter()
+
+    def run(self):
+        download_file(self.url, self.workdir.joinpath(self.outfile))
+        self.mark_complete()
+
+    @property
+    def stage_number(self) -> int:
+        return 0
+
+
+class ExtractArchive(WorkTask):
+    """
+    Extracts the downloaded file in the workdir(optionally in subdir inside
+    workdir)
+
+    Parameter
+        infile: filename which has to be extracted from the
+            download task working directory
+        download(DownloadCorpus): task which downloads the corpus to be
+            extracted
+    Requires:
+        download(DownloadCorpus): task which downloads the corpus to be
+            extracted
+    """
+
+    infile = luigi.Parameter()
+    download = luigi.TaskParameter(
+        visibility=luigi.parameter.ParameterVisibility.PRIVATE
+    )
+    # Outdir is the sub dir inside the workdir to extract the file.
+    # If set to None the file is extracted in the workdir without any
+    # subdir
+    outdir = luigi.Parameter(default=None)
+
+    def requires(self):
+        return {"download": self.download}
+
+    def run(self):
+        archive_path = self.requires()["download"].workdir.joinpath(self.infile)
+        archive_path = archive_path.absolute()
+        output_path = self.workdir
+        if self.outdir is not None:
+            output_path = output_path.joinpath(self.outdir)
+        shutil.unpack_archive(archive_path, output_path)
+
+        self.mark_complete()
 
 
 def get_download_and_extract_tasks(config: Dict):
@@ -19,10 +84,8 @@ def get_download_and_extract_tasks(config: Dict):
     tasks = {}
     for name, url in config["download_urls"].items():
         filename = os.path.basename(urlparse(url).path)
-        task = luigi_util.ExtractArchive(
-            download=luigi_util.DownloadCorpus(
-                url=url, outfile=filename, data_config=config
-            ),
+        task = ExtractArchive(
+            download=DownloadCorpus(url=url, outfile=filename, data_config=config),
             infile=filename,
             outdir=name,
             data_config=config,
@@ -32,7 +95,7 @@ def get_download_and_extract_tasks(config: Dict):
     return tasks
 
 
-class ExtractMetadata(luigi_util.WorkTask):
+class ExtractMetadata(WorkTask):
     """
     This is an abstract class that extracts metadata over the full dataset.
 
@@ -64,7 +127,7 @@ class ExtractMetadata(luigi_util.WorkTask):
         # return { "train": self.train, "test": self.test }
 
     @staticmethod
-    def slugify_file_name(relative_path: str) -> str:
+    def slugify_file_name(relative_path: str):
         """
         This is the filename in our dataset.
 
@@ -77,7 +140,7 @@ class ExtractMetadata(luigi_util.WorkTask):
         """
         name, ext = os.path.splitext(os.path.basename(relative_path))
         return f"{slugify(str(name))}"
-    
+
     @staticmethod
     def get_subsample_key(slug: str):
         """
@@ -101,7 +164,7 @@ class ExtractMetadata(luigi_util.WorkTask):
         to return a tuple as well
         """
         # Filename hash is a unique hash at a file level.
-        filename_hash = luigi_util.filename_to_int_hash(slug)
+        filename_hash = filename_to_int_hash(slug)
         # This way the sampling will be at file level
         # To make some grouped subsampling please consider overriding this
         # and returning a tuple (see speech_command.py for example)
@@ -141,36 +204,87 @@ class ExtractMetadata(luigi_util.WorkTask):
             raise ValueError("%s task_type unknown" % self.data_config["task_type"])
 
         process_metadata.to_csv(
-            os.path.join(self.workdir, self.outfile),
+            self.workdir.joinpath(self.outfile),
             index=False,
         )
 
         self.mark_complete()
 
 
-class SubsampleSplit(luigi_util.SubsampleSplit):
+class SubsampleSplit(WorkTask):
     """
     A subsampler that acts on a specific split.
     All instances of this will depend on the combined process metadata csv.
+
+    Parameters:
+        split: name of the split for which subsampling has to be done
+        max_files: maximum files required from the subsampling
+        metadata (ExtractMetadata): task which extracts a corpus level metadata
+    Requirements:
+        metadata (ExtractMetadata): task which extracts a corpus level metadata
     """
 
+    split = luigi.Parameter()
+    max_files = luigi.IntParameter()
     metadata = luigi.TaskParameter()
 
     def requires(self):
         # The meta files contain the path of the files in the data
         # so we dont need to pass the extract as a dependency here.
         return {
-            "meta": self.metadata,
+            "metadata": self.metadata,
         }
 
+    def get_metadata(self):
+        metadata = pd.read_csv(
+            self.requires()["metadata"].workdir.joinpath(
+                self.requires()["metadata"].outfile
+            )
+        )[["subsample_key", "slug", "relpath", "split"]]
 
-class SubsampleSplits(luigi_util.WorkTask):
+        # Since event detection metadata will have duplicates, we de-dup
+        # TODO: We might consider different choices of subset
+        metadata = (
+            metadata.sort_values(by="subsample_key")
+            # Drop duplicates as the subsample key is expected to be unique
+            .drop_duplicates(subset="subsample_key", ignore_index=True)
+            # Select the split to subsample
+            .loc[lambda df: df["split"] == self.split]
+        )
+        return metadata
+
+    def run(self):
+
+        metadata = self.get_metadata()
+        num_files = len(metadata)
+        max_files = num_files if self.max_files is None else self.max_files
+        if num_files > max_files:
+            print(
+                f"{len(str(num_files))} audio files in corpus, keeping only {max_files}"
+            )
+
+        # Sort by the subsample key and select the max_files number of samples
+        metadata = metadata.sort_values(by="subsample_key").iloc[:max_files]
+
+        for _, audio in metadata.iterrows():
+            audiofile = Path(audio["relpath"])
+            newaudiofile = Path(
+                self.workdir.joinpath(f"{audio['slug']}{audiofile.suffix}")
+            )
+            newaudiofile.unlink(missing_ok=True)
+            newaudiofile.symlink_to(audiofile.resolve())
+
+        self.mark_complete()
+
+
+class SubsampleSplits(WorkTask):
     """
     Aggregates subsampling of all the splits into a single task as dependencies.
-    All the subsampled files are stored in the requires workdir, so we just link to
-    that since there aren't any real outputs associated with this task.
-    This is a bit of a hack -- but it allows us to avoid rewriting
-    the Subsample task as well as take advantage of Luigi concurrency.
+
+    Parameter:
+        metadata (ExtractMetadata): task which extracts a corpus level metadata
+    Requires:
+        subsample_splits (list(SubsampleSplit)): task subsamples each split
     """
 
     metadata = luigi.TaskParameter()
@@ -198,7 +312,16 @@ class SubsampleSplits(luigi_util.WorkTask):
         self.mark_complete()
 
 
-class MonoWavTrimCorpus(luigi_util.MonoWavTrimCorpus):
+class MonoWavTrimCorpus(WorkTask):
+    """
+    Converts the file to mono, changes to wav encoding,
+    trims and pads the audio to be same length
+
+    Parameters
+        metadata (ExtractMetadata): task which extracts a corpus level metadata
+    Requires:
+        corpus(SubsampleSplits): task which aggregates the subsampling for each split
+    """
 
     metadata = luigi.TaskParameter()
 
@@ -209,8 +332,22 @@ class MonoWavTrimCorpus(luigi_util.MonoWavTrimCorpus):
             )
         }
 
+    def run(self):
+        # TODO: this should check to see if the audio is already a mono wav at the
+        #   correct length and just create a symlink if that is this case.
+        for audiofile in tqdm(list(glob(f"{self.requires()['corpus'].workdir}/*"))):
 
-class SplitTrainTestCorpus(luigi_util.SplitTrainTestCorpus):
+            newaudiofile = new_basedir(
+                os.path.splitext(audiofile)[0] + ".wav", self.workdir
+            )
+            audio_util.mono_wav_and_fix_duration(
+                audiofile, newaudiofile, duration=self.data_config["sample_duration"]
+            )
+
+        self.mark_complete()
+
+
+class SplitTrainTestCorpus(WorkTask):
 
     metadata = luigi.TaskParameter()
 
@@ -221,55 +358,191 @@ class SplitTrainTestCorpus(luigi_util.SplitTrainTestCorpus):
             "corpus": MonoWavTrimCorpus(
                 metadata=self.metadata, data_config=self.data_config
             ),
-            "meta": self.metadata,
+            "metadata": self.metadata,
         }
 
+    def run(self):
+        # Get the process metadata. This gives the freedom of picking the train test
+        # label either from the provide metadata file or any other method.
 
-class SplitTrainTestMetadata(luigi_util.SplitTrainTestMetadata):
+        # Writing slug and split makes it explicit that these columns are required
+        meta = self.requires()["metadata"]
+        process_metadata = pd.read_csv(
+            os.path.join(meta.workdir, meta.outfile),
+        )[["slug", "split"]]
+
+        # Go over the subsampled folder and pick the audio files. The audio files are
+        # saved with their slug names and hence the corresponding label can be picked
+        # up from the preprocess config
+        for audiofile in tqdm(self.requires()["corpus"].workdir.glob("*.wav")):
+            # The metadata slug doesnot have the extension as monowavtrim has not
+            # been run yet so get the name without the extension
+            split = process_metadata.loc[
+                process_metadata["slug"] == audiofile.stem, "split"
+            ].values[0]
+            split_dir = self.workdir.joinpath(split)
+            split_dir.mkdir(exist_ok=True)
+            newaudiofile = new_basedir(audiofile, split_dir)
+            os.symlink(os.path.realpath(audiofile), newaudiofile)
+
+        self.mark_complete()
+
+
+class SplitTrainTestMetadata(WorkTask):
+    """
+    Parameters
+        metadata (ExtractMetadata): task which extracts a corpus level metadata
+    Requires
+        traintestcorpus(SplitTrainTestCorpus): which produces the split
+            level corpus
+    """
 
     metadata = luigi.TaskParameter()
 
     def requires(self):
-        # Requires the traintestcorpus and the metadata.
-        # The metadata is split into train and test files
-        # which are in the traintestcorpus
         return {
             "traintestcorpus": SplitTrainTestCorpus(
                 metadata=self.metadata, data_config=self.data_config
             ),
-            "meta": self.metadata,
+            "metadata": self.metadata,
         }
 
+    def get_metadata(self):
+        metadata = pd.read_csv(
+            self.requires()["metadata"].workdir.joinpath(
+                self.requires()["metadata"].outfile
+            )
+        )
+        return metadata
 
-class MetadataVocabulary(luigi_util.MetadataVocabulary):
+    def run(self):
+        labeldf = self.get_metadata()
+
+        for split_path in self.requires()["traintestcorpus"].workdir.iterdir():
+            audiodf = pd.DataFrame(
+                [(a.stem, a.suffix) for a in split_path.glob("*.wav")],
+                columns=["slug", "ext"],
+            )
+            assert len(audiodf) != 0, f"No audio files found in: {split_path}"
+            assert (
+                not audiodf["slug"].duplicated().any()
+            ), "Duplicate files in: {split_path}"
+
+            # Get the label from the metadata with the help of the slug of the filename
+            sublabeldf = (
+                labeldf.merge(audiodf, on="slug")
+                .assign(slug=lambda df: df["slug"] + df["ext"])
+                .drop("ext", axis=1)
+            )
+
+            if self.data_config["task_type"] == "scene_labeling":
+                # Check if all the labels were found from the metadata
+                assert len(sublabeldf) == len(audiodf)
+            elif self.data_config["task_type"] == "event_labeling":
+                # This won't work for sound event detection where there might be
+                # zero or more than one event per file
+                pass
+            else:
+                raise ValueError("Invalid task_type in dataset config")
+
+            # Save the slug and the label in as the split metadata
+            sublabeldf.to_csv(
+                self.workdir.joinpath(f"{split_path.stem}.csv"),
+                index=False,
+            )
+
+        self.mark_complete()
+
+
+class MetadataVocabulary(WorkTask):
+    """
+    Parameters
+        metadata (ExtractMetadata): task which extracts a corpus level metadata
+    Requires
+        traintestmeta(SplitTrainTestMetadata): task which produces the split
+            level metadata
+    """
 
     metadata = luigi.TaskParameter()
 
     def requires(self):
-        # Depends only on the train test metadata
         return {
             "traintestmeta": SplitTrainTestMetadata(
                 metadata=self.metadata, data_config=self.data_config
             )
         }
 
+    def run(self):
+        labelset = set()
+        # Iterate over all the files in the traintestmeta and get the split_metadata
+        for split_metadata in self.requires()["traintestmeta"].workdir.iterdir():
+            labeldf = pd.read_csv(split_metadata)
+            labelset = labelset | set(labeldf["label"].unique().tolist())
 
-class ResampleSubCorpus(luigi_util.ResampleSubCorpus):
+        # Build the label idx csv and save it
+        labelcsv = pd.DataFrame(
+            [(label, idx) for (idx, label) in enumerate(sorted(list(labelset)))],
+            columns=["idx", "label"],
+        )
 
+        labelcsv.to_csv(
+            os.path.join(self.workdir, "labelvocabulary.csv"),
+            columns=["idx", "label"],
+            index=False,
+        )
+
+        self.mark_complete()
+
+
+class ResampleSubCorpus(WorkTask):
+    """
+    Resamples the Subsampled corpus in different sampling rate
+    Parameters
+        split(str): The split for which the resampling has to be done
+        sr(int): output sampling rate
+        metadata (ExtractMetadata): task which extracts a corpus level metadata
+    Requires
+        traintestcorpus(SplitTrainTestCorpus): task which produces the split
+            level corpus
+    """
+
+    sr = luigi.IntParameter()
+    split = luigi.Parameter()
     metadata = luigi.TaskParameter()
 
     def requires(self):
-        # Requires the train test corpus and will take in
-        # parameter for which split and sr the resampling
-        # has to be done
         return {
             "traintestcorpus": SplitTrainTestCorpus(
                 metadata=self.metadata, data_config=self.data_config
             )
         }
 
+    def run(self):
+        original_dir = self.requires()["traintestcorpus"].workdir.joinpath(
+            str(self.split)
+        )
+        resample_dir = self.workdir.joinpath(str(self.sr)).joinpath(str(self.split))
+        resample_dir.mkdir(parents=True, exist_ok=True)
+        for audiofile in tqdm(original_dir.glob("*.wav")):
+            resampled_audiofile = new_basedir(audiofile, resample_dir)
+            audio_util.resample_wav(audiofile, resampled_audiofile, self.sr)
 
-class FinalizeCorpus(luigi_util.FinalizeCorpus):
+        self.mark_complete()
+
+
+class FinalizeCorpus(WorkTask):
+    """
+    Create a final corpus, no longer in _workdir but in the top-level
+    at directory config.TASKNAME.
+    Parameters:
+        sample_rates (list(int)): The list of sampling rates in which the corpus
+            is required
+        metadata (ExtractMetadata): task which extracts a corpus level metadata
+    Requires:
+        resample(List(ResampleSubCorpus)): list of task which resamples each split
+        traintestmeta(SplitTrainTestMetadata): task which produces the split
+            level metadata
+    """
 
     sample_rates = luigi.ListParameter()
     metadata = luigi.TaskParameter()
@@ -296,6 +569,42 @@ class FinalizeCorpus(luigi_util.FinalizeCorpus):
             ),
         }
 
+    # We overwrite workdir here, because we want the output to be
+    # the finalized top-level task directory
+    @property
+    def workdir(self):
+        return Path("tasks").joinpath(self.versioned_task_name)
+
+    def run(self):
+        if self.workdir.exists():
+            shutil.rmtree(self.workdir)
+
+        # Copy the resampled files
+        # The workdirectory of the all the resample task is same
+        # so select the first one
+        shutil.copytree(self.requires()["resample"][0].workdir, self.workdir)
+
+        # Copy the traintestmetadata
+        shutil.copytree(
+            self.requires()["traintestmeta"].workdir,
+            self.workdir,
+            dirs_exist_ok=True,
+        )
+        # Copy the vocabmetadata
+        shutil.copytree(
+            self.requires()["vocabmeta"].workdir,
+            self.workdir,
+            dirs_exist_ok=True,
+        )
+        # Save the dataset config as a json file
+        config_out = self.workdir.joinpath("dataset_metadata.json")
+        with open(config_out, "w") as fp:
+            json.dump(
+                self.data_config, fp, indent=True, cls=luigi.parameter._DictParamEncoder
+            )
+
+        self.mark_complete()
+
 
 def run(task: Union[List[luigi.Task], luigi.Task], num_workers: int):
     """
@@ -310,7 +619,7 @@ def run(task: Union[List[luigi.Task], luigi.Task], num_workers: int):
     if isinstance(task, luigi.Task):
         task = [task]
 
-    luigi_util.ensure_dir("_workdir")
+    Path("_workdir").mkdir(exist_ok=True)
     luigi.build(
         task,
         workers=num_workers,
